Refactor: AuthController 기능 UserController로 이동
