+++ src/main/java/com/sb02/blogpractice/controller/UserController.java
@@ -1,7 +1,11 @@
 package com.sb02.blogpractice.controller;
 
 import com.sb02.blogpractice.dto.CreateRequestUserDTO;
+import com.sb02.blogpractice.dto.LoginParam;
+import com.sb02.blogpractice.dto.LoginResponseDTO;
 import com.sb02.blogpractice.dto.UserDTO;
+import com.sb02.blogpractice.jwt.JwtUtil;
+import com.sb02.blogpractice.service.AuthService;
 import com.sb02.blogpractice.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +20,16 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/users")
 public class UserController {
     private final UserService userService;
+    private final AuthService authService;
+    private final JwtUtil jwtUtil;
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponseDTO> login(@RequestBody LoginParam loginParam) {
+        UserDTO userDTO =  authService.login(loginParam);
+        String accessToken = jwtUtil.generateToken(userDTO.id());
+        LoginResponseDTO loginResponseDTO = new LoginResponseDTO(userDTO, accessToken);
+        return ResponseEntity.ok(loginResponseDTO);
+    }
 
     @PostMapping("/register")
     public ResponseEntity<UserDTO> register(@Valid @RequestBody CreateRequestUserDTO createRequestUserDTO) {
